iot snapshot rate limiter
