+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/storageengine/rescon/quotas/QuotaLimiter.java
@@ -22,6 +22,9 @@ package org.apache.iotdb.db.storageengine.rescon.quotas;
 import org.apache.iotdb.common.rpc.thrift.TTimedQuota;
 import org.apache.iotdb.common.rpc.thrift.ThrottleType;
 import org.apache.iotdb.commons.exception.RpcThrottlingException;
+import org.apache.iotdb.commons.quotas.AverageIntervalRateLimiter;
+import org.apache.iotdb.commons.quotas.FixedIntervalRateLimiter;
+import org.apache.iotdb.commons.quotas.RateLimiter;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.rpc.TSStatusCode;
